# remove passthrough and lazily load headers