Replaced passthrough stream with just using ReadableStream instead and
caching the controller instead of a WritableStream from a passthrough.

Also lazily construct the Headers instances on Request only when needed.

Together this results in about a 25% boost in requests per second and a
45% reduction in latency for our hello-world benchmark.

---

Here are the results on my machine.

Before change:

```
$ wrk -d 10 -t 10 http://localhost:3000/hello
Running 10s test @ http://localhost:3000/hello
  10 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   769.43us    1.55ms  19.27ms   89.77%
    Req/Sec     3.78k   570.99     5.13k    80.10%
  376931 requests in 10.02s, 41.34MB read
Requests/sec:  37602.11
Transfer/sec:      4.12MB
```

After change:

```
$ wrk -d 10 -t 10 http://localhost:3000/hello
Running 10s test @ http://localhost:3000/hello
  10 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   504.32us    1.11ms  21.43ms   92.28%
    Req/Sec     4.80k   584.47    10.94k    84.93%
  479198 requests in 10.10s, 52.55MB read
Requests/sec:  47436.55
Transfer/sec:      5.20MB
```