@tlhunter 2019-07-16
56k rps:
```
$ wrk -c 100 -d 60 http://localhost:3000/hello
Running 1m test @ http://localhost:3000/hello
  2 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     3.07ms    6.76ms 130.72ms   95.42%
    Req/Sec    28.49k     4.53k   37.01k    82.05%
  3402137 requests in 1.00m, 373.12MB read
Requests/sec:  56608.64
Transfer/sec:      6.21MB
```