@tlhunter 2019-07-16
51k rps after bugfix:

```
$ wrk -c 100 -d 60 http://localhost:3000/hello
Running 1m test @ http://localhost:3000/hello
  2 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     3.12ms    6.34ms 114.79ms   95.38%
    Req/Sec    25.79k     3.89k   33.79k    81.89%
  3077517 requests in 1.00m, 337.52MB read
Requests/sec:  51278.65
Transfer/sec:      5.62MB
```